add reservation model
