+++ SISTWebProject-1/WebContent/css/style.css
@@ -151,6 +151,6 @@
 	font-size: 18px;
 	border: none;
 	display: block;
-	width: 100%;
+	width: 50%;
 	text-transform: capitalize;
 }
