+++ SISTWebProject-1/src/com/sist/reservation/model/ReservationModel.java
@@ -19,11 +19,25 @@ public class ReservationModel {
 	@RequestMapping("reservation/reservation_ok.do")
 	public String reservation_ok(HttpServletRequest request, HttpServletResponse response)
 	{
-		String no=request.getParameter("no");
-		ReservationVO vo=ReservationDAO.reservationListData(Integer.parseInt(no));
-		request.setAttribute("vo", vo);
-		request.setAttribute("main_jsp","???");
-		return "????"; //어디로 갈지 생각해보기.....
+		try
+		{
+			request.setCharacterEncoding("UTF-8");
+		}catch(Exception ex){}
+		String name=request.getParameter("name");
+		String email=request.getParameter("email");
+		String date=request.getParameter("date");
+		String time=request.getParameter("time");
+		String person=request.getParameter("person");
+		String tel=request.getParameter("tel");
+		
+		ReservationVO rvo=new ReservationVO();
+		//rvo.setMemberid(memberid); 								-- from session isLogin
+		//rvo.setBookingnumber(bookingnumber);						-- unused variables / TO_DATE(insert)
+		
+		rvo.setTime(time);
+		rvo.setInwon(Integer.parseInt(person));
+		rvo.setReservationdate(date);
+		return "redirect:../reservation/reservation_ok.do";
 	}
 }
